changed the name of the workflow
